Pass the task_struct explicitly to delayacct_blkio_end
In this code path, `current` refers to the process which was previously
running (and has been interrupted) on the CPU. We want to update
statistics for the task whose I/O just finished. For that reason, we'll
need to pass the task_struct of that process into delayacct_blkio_end.